Separate ProjectRazorJson version between Json and MessagePack
