+++ src/Razor/src/Microsoft.AspNetCore.Razor.ProjectEngineHost/ProjectSystem/ProjectRazorJson.cs
@@ -9,12 +9,6 @@ namespace Microsoft.AspNetCore.Razor.ProjectSystem;
 
 internal sealed class ProjectRazorJson
 {
-    // This version number must be incremented if the serialization format for ProjectRazorJson
-    // or any of the types that compose it changes. This includes: RazorConfiguration,
-    // ProjectWorkspaceState, TagHelperDescriptor, and DocumentSnapshotHandle.
-	// NOTE: If this version is changed, a coordinated insertion is required between Roslyn and Razor for the C# extension.
-    public const int Version = 2;
-
     public string SerializedFilePath { get; }
     public string FilePath { get; }
     public RazorConfiguration? Configuration { get; }
